% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Schumaker.R
\name{Schumaker}
\alias{Schumaker}
\title{Create a Schumaker Spline}
\usage{
Schumaker(tt, FF, ff = "Not-Supplied", Vectorised = FALSE,
  Extrapolation = c("Curve", "Constant", "Linear"))
}
\arguments{
\item{tt}{A vector of x coordinates}

\item{FF}{A corresponding vector of y coordinates}

\item{ff}{(Optional) A corresponding vector of gradiants at the data points. If not supplied this is estimated.}

\item{Vectorised}{This is a boolean parameter. Set to TRUE if you want to be able to input vectors to the created spline. If you will only input single values set this to FALSE as it is a bit faster.}

\item{Extrapolation}{This determines how the spline function responds when an input is recieved outside the domain of tt. The options are "Curve" which outputs the result of the point on the quadratic curve at the nearest interval, "Constant" which outputs the FF value at the end of the tt domain and "Linear" which extends the spline using the gradiant at the edge of tt.}
}
\value{
A list with 3 spline functions. Thee first spline is is for the input points, the second spline is the first derivative of the first spline, the third spline is the second derivative. Each function takes an x value (or vector if Vectorised = TRUE) and outputs the interpolated y value (or relevent derivative).
}
\description{
Create a Schumaker Spline
}
\examples{
tt = c(0,1,2,3,4,5,6)
FF = c(2,1, 0.59, 0.27, 0.25, -0.23, -0.45)

SSS = Schumaker(tt,FF, Vectorised = TRUE)
Spline   = SSS[[1]]
SplineD  = SSS[[2]]
Spline2D = SSS[[3]]

xarray = seq(0,6,0.1)
Res = Spline(xarray)
Res2 = SplineD(xarray)
Res3 = Spline2D(xarray)

plot(xarray, Res, ylim=c(-2,5))
lines(xarray, Res2, col = 2)
lines(xarray, Res3, col = 3)
}

