% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SchumakerIndInterval.R
\name{SchumakerIndInterval}
\alias{SchumakerIndInterval}
\title{SchumakerIndInterval}
\usage{
SchumakerIndInterval(z, s, Smallt)
}
\arguments{
\item{z}{This is the y value at edges of an interval.}

\item{s}{This is the slope at edges of an interval.}

\item{Smallt}{This is x values at the edge of an interval.}
}
\value{
The location of the knot and quadratic coefficients for an interval.
}
\description{
This creates quadratic coefficients for one interval of a domain.
This is an internal function that is called from the Schumaker function.
}

